requirements.txt
paho-mqtt>=1.6.1
pyseneye>=0.1.7
